4.1.5 Отрефакторила модуль с утилитами
