+++ src/mock/point.js
@@ -1,4 +1,4 @@
-import { getRandomInteger } from '../utils.js';
+import { getRandomInteger } from '../utils/common.js';
 import dayjs from 'dayjs';
 import { TYPES } from '../const.js';
 import { generateDestination } from './destination.js';
