+++ src/utils/point.js
@@ -1,32 +1,14 @@
 import dayjs from 'dayjs';
 
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
+const humanizePointDate = (date) => date ? dayjs(date).format('MMM DD') : '';
+const humanizePointDateForAttribute = (date) => date ? dayjs(date).format('YYYY-MM-DD') : '';
 
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
+const humanizePointTime = (date) => date ? dayjs(date).format('HH:mm') : '';
+const humanizePointTimeForAttribute = (date) => date ? dayjs(date).format('YYYY-MM-DDTHH:mm') : '';
 
-const humanizePointDate = (date) => dayjs(date).format('MMM DD');
-const humanizePointDateForAttribute = (date) => dayjs(date).format('YYYY-MM-DD');
+const humanizePointDateTime = (date) => date ? dayjs(date).format('DD/MM/YY HH:mm') : '';
 
-const humanizePointTime = (date) => dayjs(date).format('HH:mm');
-const humanizePointTimeForAttribute = (date) => dayjs(date).format('YYYY-MM-DDTHH:mm');
-
-const humanizePointDateTime = (dateFrom) => dayjs(dateFrom).format('DD/MM/YY HH:mm');
-
-const humanizeTripDates = (dateFrom, dateTo) => {
-  const startMonth = dayjs(dateFrom).format('MMM');
-  const endMonth = dayjs(dateTo).format('MMM');
-
-  if (startMonth === endMonth) {
-    return `${dayjs(dateFrom).format('MMM DD')} &nbsp;&mdash;&nbsp; ${dayjs(dateTo).format('DD')}`;
-  }
-
-  return `${dayjs(dateFrom).format('MMM DD')} &nbsp;&mdash;&nbsp; ${dayjs(dateTo).format('MMM DD')}`;
-};
-
-const getDuration = (dateTo, dateFrom) => dayjs(dateTo).diff(dayjs(dateFrom));
+const getDuration = (dateTo, dateFrom) => dateTo && dateFrom ? dayjs(dateTo).diff(dayjs(dateFrom)) : '';
 
 const millisecondsPerSecond = 1000;
 const secondsPerMinute = 60;
@@ -66,13 +48,11 @@ const humanizeDuration = (duration) => {
 };
 
 export {
-  getRandomInteger,
   humanizePointDate,
   humanizePointDateForAttribute,
   humanizePointTime,
   humanizePointTimeForAttribute,
   humanizePointDateTime,
-  humanizeTripDates,
   getDuration,
   humanizeDuration
 };
